back/requirements.txt
Django                        
django-cors-headers           
djangorestframework           
djangorestframework-simplejwt 
pillow 
psycopg2-binary
PyJWT
sqlparse